requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
Pillow==10.4.0
pyjwt==2.9.0
torch
torchvision
torchaudio
diffusers==0.30.0
transformers==4.44.2
accelerate==0.34.2
safetensors==0.4.4